# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Katha API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/rewards/checkin": {
      "get": {
        "tags": [
          "rewards"
        ],
        "summary": "Checkin Status",
        "operationId": "checkin_status_v1_rewards_checkin_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "rewards"
        ],
        "summary": "Checkin",
        "operationId": "checkin_v1_rewards_checkin_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/rewards/tasks": {
      "get": {
        "tags": [
          "rewards"
        ],
        "summary": "Tasks",
        "operationId": "tasks_v1_rewards_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-katha-platform",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Katha-Platform"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskOut"
                  },
                  "title": "Response Tasks V1 Rewards Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rewards/tasks/{task_id}/claim": {
      "post": {
        "tags": [
          "rewards"
        ],
        "summary": "Claim",
        "operationId": "claim_v1_rewards_tasks__task_id__claim_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/translations/{lang}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Translations",
        "description": "UI strings for a language, falling back to English key by key.",
        "operationId": "translations_v1_translations__lang__get",
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pages": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Footer Pages",
        "operationId": "footer_pages_v1_pages_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FooterLink"
                  },
                  "title": "Response Footer Pages V1 Pages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pages/{slug}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Page",
        "operationId": "page_v1_pages__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CmsPageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinOut": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "streak_day": {
            "type": "integer",
            "title": "Streak Day"
          },
          "coins": {
            "type": "integer",
            "title": "Coins"
          },
          "coin_balance": {
            "type": "integer",
            "title": "Coin Balance"
          }
        },
        "type": "object",
        "required": [
          "day",
          "streak_day",
          "coins",
          "coin_balance"
        ],
        "title": "CheckinOut"
      },
      "CheckinStatus": {
        "properties": {
          "checked_in_today": {
            "type": "boolean",
            "title": "Checked In Today"
          },
          "streak_day": {
            "type": "integer",
            "title": "Streak Day"
          },
          "next_streak_day": {
            "type": "integer",
            "title": "Next Streak Day"
          },
          "rewards": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Rewards"
          },
          "coin_balance": {
            "type": "integer",
            "title": "Coin Balance"
          }
        },
        "type": "object",
        "required": [
          "checked_in_today",
          "streak_day",
          "next_streak_day",
          "rewards",
          "coin_balance"
        ],
        "title": "CheckinStatus"
      },
...
      "ClaimOut": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "coins": {
            "type": "integer",
            "title": "Coins"
          },
          "coin_balance": {
            "type": "integer",
            "title": "Coin Balance"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "coins",
          "coin_balance"
        ],
        "title": "ClaimOut"
      },
      "ClaimRequest": {
        "properties": {
          "ad_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Event Id"
          }
        },
        "type": "object",
        "title": "ClaimRequest"
      },
...
      "CmsPageOut": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body_html": {
            "type": "string",
            "title": "Body Html"
          },
          "lang": {
            "type": "string",
            "title": "Lang"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "body_html",
          "lang"
        ],
        "title": "CmsPageOut"
      },
...
      "FooterLink": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title"
        ],
        "title": "FooterLink"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RewardFrequency": {
        "type": "string",
        "enum": [
          "once",
          "daily"
        ],
        "title": "RewardFrequency"
      },
...
      "RewardTaskKind": {
        "type": "string",
        "enum": [
          "link",
          "rewarded_ad",
          "share",
          "follow"
        ],
        "title": "RewardTaskKind"
      },
...
      "TaskOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/RewardTaskKind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "coins": {
            "type": "integer",
            "title": "Coins"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "timer_seconds": {
            "type": "integer",
            "title": "Timer Seconds"
          },
          "frequency": {
            "$ref": "#/components/schemas/RewardFrequency"
          },
          "claimed": {
            "type": "boolean",
            "title": "Claimed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "description",
          "coins",
          "url",
          "timer_seconds",
          "frequency",
          "claimed"
        ],
        "title": "TaskOut"
      },
...
      "TranslationsOut": {
        "properties": {
          "lang": {
            "type": "string",
            "title": "Lang"
          },
          "messages": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "lang",
          "messages"
        ],
        "title": "TranslationsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}